completed basket

+++ server/models/models.js
@@ -34,7 +34,7 @@ const Device = sequelize.define('device', {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     name: { type: DataTypes.STRING, unique: true, allowNull: false },
     price: { type: DataTypes.INTEGER, allowNull: false },
-    rating: { type: DataTypes.INTEGER, defaultValue: 0 },
+    rating: { type: DataTypes.REAL, defaultValue: 0 },
     img: { type: DataTypes.STRING, allowNull: false },
 });
